+++ apps/backoffice-v2/src/pages/Entity/components/Case/interfaces.ts
@@ -1,18 +1,19 @@
+import { ComponentProps } from 'react';
+
 import { Actions } from './Case.Actions';
 import { Documents } from './Case.Documents';
 import { Info } from './Case.Info';
 import { Content } from './Case.Content';
 import { FaceMatch } from './Case.FaceMatch';
 import { TState } from '../../../../common/types';
-import { ComponentProps } from 'react';
+import { Assignee } from '../../../../common/components/atoms/AssignDropdown/AssignDropdown';
 
 export interface IItemProps {
   id: string;
   fullName: string;
   createdAt: string;
   avatarUrl: string;
-  assigneeId: string;
-  assigneeFullName: string;
+  assignee: Assignee;
   status: TState;
 }