+++ services/workflows-service/src/providers/file/file-provider/local-file.service.ts
@@ -9,43 +9,40 @@ export class LocalFileService implements IFileProvider {
     this.client = fs;
   }
 
-  async downloadFile(
+  async download(
     remoteFileConfig: TRemoteFileConfig,
     localFilePath: TLocalFilePath,
   ): Promise<TLocalFilePath> {
-    return await this.copyFile(remoteFileConfig as TRemoteUri, localFilePath);
+    return await this.copy(remoteFileConfig as TRemoteUri, localFilePath);
   }
 
   // eslint-disable-next-line @typescript-eslint/require-await
-  async isRemoteFileExists(remoteFileConfig: TLocalFilePath): Promise<boolean> {
+  async isRemoteExists(remoteFileConfig: TLocalFilePath): Promise<boolean> {
     const localFilePath = remoteFileConfig;
 
     return this.client.existsSync(localFilePath);
   }
 
-  async copyFile(
-    fromFilePath: TLocalFilePath,
-    toFilePath: TLocalFilePath,
-  ): Promise<TLocalFilePath> {
+  async copy(fromFilePath: TLocalFilePath, toFilePath: TLocalFilePath): Promise<TLocalFilePath> {
     // eslint-disable-next-line @typescript-eslint/await-thenable
     await this.client.copyFileSync(fromFilePath, toFilePath);
 
     return toFilePath;
   }
 
-  async uploadFile(
+  async upload(
     localFilePath: TLocalFilePath,
     remoteFileConfig: TRemoteFileConfig,
   ): Promise<TRemoteFileConfig> {
-    const toLocalFilePAth = remoteFileConfig as TLocalFilePath;
+    const toLocalFilePath = remoteFileConfig as TLocalFilePath;
     this.client
       .createReadStream(localFilePath)
-      .pipe(this.client.createWriteStream(toLocalFilePAth));
+      .pipe(this.client.createWriteStream(toLocalFilePath));
 
-    return Promise.resolve(toLocalFilePAth);
+    return Promise.resolve(toLocalFilePath);
   }
 
-  generateRemoteFilePath({
+  generateRemotePath({
     fileName,
     customerName,
     directory,