+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ballerine/blocks
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.2
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.1
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.0
+
 ## 0.1.0
 
 ### Minor Changes