+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,26 @@
 # web-ui-sdk
 
+## 1.4.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.2
+
+## 1.4.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.1
+
+## 1.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.0
+
 ## 1.4.0
 
 ### Minor Changes