+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/workflow-node-sdk
 
+## 0.5.3
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.3
+
+## 0.5.2
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.2
+
+## 0.5.1
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes